auth/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
bson==0.5.7
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
csiphash==0.0.5
Deprecated==1.2.13
Flask==2.1.3
Flask-Dance==6.0.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lazy==1.4
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
peewee==3.15.1
prompt-toolkit==3.0.30
pycparser==2.21
PyJWT==1.6.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rfc7539==1.2.1
six==1.16.0
ua-parser==0.15.0
urllib3==1.26.11
URLObject==2.4.3
user-agents==1.1.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.1
wrapt==1.14.1
zipp==3.8.1
